cloudbuild: use --load flag to allow pushing images
The cloud build uses the `docker-container` driver which does not make
build images available in the local image store by default -- only in
the cache. This means the cloud build process doesn't see the images in
order to push them at the end. Use the --load flag to fix this.

See: https://docs.docker.com/build/drivers/docker-container/#loading-to-local-image-store

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>